requirements.txt
pandas
mysql-connector-python
google-cloud-bigquery
google-auth
db-dtypes
notebook
seaborn
matplotlib
